python/LinearRegressionModelTrainingService/requirements.txt
requests
pydantic
scikit-learn
pandas
numpy
python-dotenv
skl2onnx
onnx
azure-storage-blob
azure-identity
pytest
pytz
apscheduler

